# openapi/imaginarium-v1.yaml
openapi: 3.1.0
info:
  title: Imaginarium-RS Local API
  version: 0.1.0
  description: |
    LAN-facing gateway for the xAI Imagine image/video API.

    **Auth.** Every `/v1/*` route requires a token; `/health` is public. A token may
    be presented three ways (any one):
      - `Authorization: Bearer <token>`
      - `X-Imaginarium-Token: <token>`
      - `?token=<token>` query parameter (for browsers / `<img>`/`<video>` sources)
    Token scopes: `read` (GET + job wait), `write` (create jobs, import, craft),
    `admin` (token management + everything). The env/config node token is admin.
    The upstream xAI key stays on the node and is never returned by the API.

    **Shipped vs planned.** Every path below is implemented. Routes that appear in
    `docs/ARCHITECTURE.md` but are NOT yet implemented — SSE job events
    (`GET /v1/jobs/{id}/events`), library listing/get/delete/upload
    (`GET/DELETE /v1/library`, `GET /v1/library/{id}`, `POST /v1/library/upload`),
    a standalone video poll route (`GET /v1/videos/{id}`), and per-token rate limits
    / spend caps — are intentionally omitted here. Do not build against them yet.
servers:
  - url: http://127.0.0.1:8791
    description: Default loopback bind. LAN use requires an explicit non-loopback bind and a token.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer <token>"
    tokenHeader:
      type: apiKey
      in: header
      name: X-Imaginarium-Token
    tokenQuery:
      type: apiKey
      in: query
      name: token
  schemas:
...
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
security:
  - bearerAuth: []
  - tokenHeader: []
  - tokenQuery: []
paths:
...
  /v1/models:
    get:
      summary: Model catalog + capability matrix
      responses:
        "200": { description: Catalog JSON }
        "401": { description: Missing/invalid token, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/estimate:
    post:
      summary: Cost estimate for an image or video job (no spend)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kind: { type: string, enum: [image, video] }
                model: { type: string, description: "concrete model, or \"auto\"/omit for the default" }
                n: { type: integer }
                duration: { type: integer }
      responses:
        "200": { description: Estimate JSON }
  /v1/images/generations:
    post:
      summary: Text-to-image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string }
                model: { type: string, description: "grok-imagine-image | grok-imagine-image-quality | grok-imagine-image-2.0 | aliases image/quality/2.0 | \"auto\" | omit" }
                n: { type: integer }
                aspect_ratio: { type: string }
                resolution: { type: string, enum: ["1k", "2k"] }
                quality: { type: string, enum: [low, medium], description: "Image 2.0 only. Omitted → upstream default medium. Rejected on other image models." }
      responses:
        "200": { description: JobResult, content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "400": { description: Invalid model/params (a bare local file path in a media field is rejected here) }
        "502": { description: Upstream xAI error }
  /v1/images/edits:
    post:
      summary: Image edit (1-3 source images)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, images]
              properties:
                prompt: { type: string }
                images:
                  type: array
                  maxItems: 3
                  items: { type: string, description: "library:{job_id} (node-library chain ref — preferred; #n suffix addresses batch asset n), data: URL, http(s) URL, or xAI file_… id. Bare local filesystem paths are rejected from the network surface." }
                model: { type: string }
                n: { type: integer }
                aspect_ratio: { type: string }
                resolution: { type: string }
                quality: { type: string, enum: [low, medium], description: "Image 2.0 only. Omitted → upstream default medium." }
      responses:
        "200": { description: JobResult, content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "400": { description: Bad request (too many/invalid images, bad model) }
        "502": { description: Upstream xAI error }
  /v1/videos/generations:
    post:
      summary: Video generate — T2V / I2V / R2V (mode inferred from fields present)
      description: "Mode inferred from fields: prompt-only (T2V), image (I2V), reference_images and/or reference_audios (R2V). image + reference_* is rejected. Default model is grok-imagine-video-1.5. 1080p is T2V/I2V only; R2V max 720p."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt: { type: string }
                model: { type: string, description: "\"auto\"/omit → grok-imagine-video-1.5 for generate modes" }
                image: { type: string, description: "I2V start frame: library:{job_id}, data:, http(s), or file_…" }
                reference_images: { type: array, items: { type: string, description: "library:{job_id}, data:, http(s), or file_…" } }
                reference_audios: { type: array, maxItems: 3, items: { type: string, description: "Video 1.5 preset voice_id (eve, ara, leo, rex, …). Tag <AUDIO_0>… in the prompt. Audio-only R2V is valid." } }
                duration: { type: integer, description: "1-15s (clamped)" }
                aspect_ratio: { type: string }
                resolution: { type: string, enum: [480p, 720p, 1080p] }
                no_wait: { type: boolean, description: "true = return after submit (poll with /v1/jobs/{id}/wait)" }
      responses:
        "200": { description: "JobResult (status may be pending/running when no_wait or on wait timeout)", content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "400": { description: Illegal mode/model/resolution combination }
        "502": { description: Upstream xAI error }
  /v1/videos/edits:
    post:
      summary: Video edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, video]
              properties:
                prompt: { type: string }
                video: { type: string, description: "library:{job_id}, data:, http(s), or file_…" }
                model: { type: string }
                no_wait: { type: boolean }
      responses:
        "200": { description: JobResult, content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "400": { description: Bad request }
        "502": { description: Upstream xAI error }
  /v1/videos/extensions:
    post:
      summary: Video extend (adds a 2-10s segment, default 6)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, video]
              properties:
                prompt: { type: string }
                video: { type: string }
                duration: { type: integer }
                model: { type: string }
                no_wait: { type: boolean }
      responses:
        "200": { description: JobResult, content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "400": { description: Bad request }
        "502": { description: Upstream xAI error }
  /v1/jobs:
    get:
      summary: List recent jobs (summary rows)
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20 } }
      responses:
        "200": { description: "Array of job summary rows: job_id, status, mode, model, created_at, error, prompt, assets (media count — address each via /content?i=N)" }
  /v1/jobs/{job_id}:
    get:
      summary: Get a job by local ULID
      parameters:
        - { name: job_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: JobResult, content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "404": { description: Unknown job_id }
  /v1/jobs/{job_id}/wait:
    post:
      summary: Block until a video job resolves (read scope)
      description: "An already-terminal job (incl. synchronous image jobs) returns immediately; an unknown id returns 404; a non-terminal video job is polled to completion or timeout."
      parameters:
        - { name: job_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: JobResult, content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "404": { description: Unknown job_id }
        "502": { description: Upstream xAI error }
  /v1/library/{job_id}/content:
    get:
      summary: Stream a media file for a job (bytes)
      description: "job_id is validated as a safe id (no path traversal). ?i=N addresses the N-th asset of a batch (default 0 — the historic first file). Content-Type is inferred from the file."
      parameters:
        - { name: job_id, in: path, required: true, schema: { type: string } }
        - { name: i, in: query, required: false, schema: { type: integer, default: 0 }, description: Asset index within the job }
      responses:
        "200": { description: Raw image/video bytes }
        "400": { description: Invalid asset id }
        "404": { description: Asset not found }
  /v1/library/{job_id}/thumb:
    get:
      summary: 480px JPEG poster for a job's first asset (thumb.jpg)
      description: "Generated eagerly on import/completion and rebuilt lazily on demand (covers pre-thumb library content). Videos take their poster ~1s in; images are downscales. Audio-only jobs 404. The thumb.jpg sidecar is never addressable as a media asset."
      parameters:
        - { name: job_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: JPEG bytes }
        "400": { description: Invalid asset id }
        "404": { description: Asset not found / no thumbnail (audio) }
  /v1/library/import:
    post:
      summary: Import bytes (data URL or base64) as a completed library job
      description: "Request body limit is 64 MB; decoded import payload is capped at 40 MB. Accepts images, video, and audio (mp3/wav/flac/ogg/m4a/opus — e.g. a Sonus track destined for a craft music bed)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data: { type: string, description: "data: URL or raw base64" }
                filename: { type: string }
                note: { type: string }
                source_job_id: { type: string }
      responses:
        "200": { description: JobResult, content: { application/json: { schema: { $ref: '#/components/schemas/JobResult' } } } }
        "400": { description: Malformed data }
        "413": { description: Decoded payload over 40 MB }
...
  /v1/tokens:
    get:
      summary: List tokens (admin) — hashes never returned
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: "{ tokens: [...] }" }
        "403": { description: Requires admin scope }
    post:
      summary: Mint a token (admin) — plaintext shown once
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
                scope: { type: string, enum: [read, write, admin], default: write }
      responses:
        "200": { description: "{ id, label, scope, token, created_at }" }
...